MicroServices/TemplateService/requirements.txt
paho-mqtt==1.6.1
SPARQLWrapper==2.0.0